Add workaround for non-working bring to front on hotkey press
A workaround has been implemented which brings the main window to front
when the hotkey is pressed and the window is already opened in the
background. The workaround has the ugly effect that the window is
shortly minimized and then reappears in the front. Unfortunately there
is no better way to do this as windows forbids applications to bring
themselves to front and "steal" focus.